requirements.txt
absl-py==2.3.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cryptography==46.0.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.116.1
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.32.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.30.0
httpx==0.28.1
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imbalanced-learn==0.13.0
imblearn==0.0
iopath==0.1.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
lightgbm==4.6.0
lime==0.2.0.1
llvmlite==0.44.0
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
motor==3.7.1
mpmath==1.3.0
narwhals==1.44.0
nest-asyncio==1.6.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
ollama==0.5.3
omegaconf==2.3.0
openai==1.102.0
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pillow==11.2.1
platformdirs==4.3.8
portalocker==3.2.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pymongo==4.14.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.4
roifile==2025.5.10
rpds-py==0.25.1
rsa==4.9.1
-e git+https://github.com/facebookresearch/sam2.git@2b90b9f5ceec907a1c18123530e92e794ad901a4#egg=SAM_2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
segment-anything==1.0
setuptools==80.9.0
shap==0.48.0
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
starlette==0.47.3
streamlit==1.46.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.6.11
toml==0.10.2
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.156
ultralytics-thop==2.0.14
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
xgboost==3.0.3
